+++ Sources/WalletConnectNotify/Client/Wallet/SubscriptionWatcher.swift
@@ -62,38 +62,4 @@ class SubscriptionWatcher {
 }
 
 
-protocol NotifyWatchSubscriptionsRequesting {
-    func watchSubscriptions() async throws
-}
 
-protocol NotificationPublishing {
-    func publisher(for name: NSNotification.Name) -> AnyPublisher<Notification, Never>
-}
-
-extension NotificationCenter: NotificationPublishing {
-    func publisher(for name: NSNotification.Name) -> AnyPublisher<Notification, Never> {
-        return publisher(for: name, object: nil).eraseToAnyPublisher()
-    }
-}
-
-
-
-class MockNotificationCenter: NotificationPublishing {
-    private let subject = PassthroughSubject<Notification, Never>()
-
-    func publisher(for name: NSNotification.Name) -> AnyPublisher<Notification, Never> {
-        return subject.eraseToAnyPublisher()
-    }
-
-    func post(name: NSNotification.Name) {
-        subject.send(Notification(name: name))
-    }
-}
-
-class MockNotifyWatchSubscriptionsRequester: NotifyWatchSubscriptionsRequesting {
-    var onWatchSubscriptions: (() -> Void)?
-
-    func watchSubscriptions() async throws {
-        onWatchSubscriptions?()
-    }
-}
